1.移动后移动范围删除 2.修复移动后移动范围位置不变的问题

+++ slg/Assets/CharacterMove/MoveRangeConponent.cs
@@ -5,9 +5,9 @@ public struct CreateMoveRangeData : IComponentData { }
 public struct DestroyMoveRangeData : IComponentData { }
 
 public struct MovePosition : IComponentData {
-    public int X;
-    public int Y;
-    public int Z;
+    public float X;
+    public float Y;
+    public float Z;
 }
